feat: implement Parry, Enemy Combat AI, Platformer Pathfinding, and Behavior Tree
- Parry/Precise-Deflect System (engine_platformer):
  * Parry component with configurable window (0.15s), stun (0.5s), cooldown (0.5s)
  * ParrySystem counts down timers each tick
  * Integrated with damageEntity: active parry negates hit, stuns attacker, emits ParrySuccessEvent
  * World.tryParry(target, attacker) extension for custom flows
  * Full JSON serialization

- Enemy Combat AI with Telegraph States (engine_platformer):
  * State machine: Idle -> Patrol -> Alert -> Approach -> Telegraph -> Attack -> Recovery -> Reposition
  * Special states: Stagger, Guard, Retreat
  * EnemyTelegraphEvent emitted on wind-up (player can parry/dodge)
  * EnemyAttackEvent emitted when hitbox activates
  * Per-entity config via AIState.memory overrides
  * Full JSON serialization

- Platformer-Aware Pathfinding (engine_platformer):
  * Extends tile-based A* with platformer movement: walk, jump, climb, fall, drop-through
  * Configurable jump distance/height, ladder support, one-way platform handling
  * Returns PlatformerPathPoint with jumpRequired, climbRequired, fallRequired, dropThroughOneWay flags
  * 8 comprehensive tests passing

- Behavior Tree / State Machine (engine_core):
  * Composite nodes: Sequence, Selector, Parallel
  * Decorators: Inverter, Repeater, Succeeder, Failer
  * Leaf nodes: BTLeaf (wraps registered Behavior), BTWait
  * Blackboard persistence via AIState.memory
  * Full JSON serialization round-trip
  * BehaviorTreeBehavior wraps BTNode for standard Behavior interface

- Fixes:
  * HearingSystem line-of-sight bug: added 2px offset to raycast start
  * Shader compilation fix: ambient_lighting.frag int -> float uniform
  * Guard break logic: apply full damage when guard breaks
  * HealthSystem stability regen blocked during guard break stun
  * Attack system cooldown logic fixed; test expectations updated

+++ packages/engine_flutter/shaders/ambient_lighting.frag
@@ -24,7 +24,7 @@
 uniform vec2 uSize;
 uniform vec4 uAmbientColor;
 uniform float uAmbientBrightness;
-uniform int uLightCount;
+uniform float uLightCount;
 uniform vec4 uLights[32];
 uniform vec4 uLightColors[32];
 uniform vec4 uLightFlags[32];
